moved otp inside magic link sign in component
